+++ vtl-spark/src/test/java/fr/insee/vtl/spark/processing/engine/analytic/AnalyticVarPopTest.java
@@ -1,4 +1,4 @@
-package fr.insee.vtl.spark.processing.engine;
+package fr.insee.vtl.spark.processing.engine.analytic;
 
 import fr.insee.vtl.engine.VtlScriptEngine;
 import fr.insee.vtl.model.Dataset;
